refactor(via): cleanup serve and rename it to accept
