+++ src/server/mod.rs
@@ -1,8 +1,8 @@
 //! Serve an [App](crate::App) over HTTP or HTTPS.
 //!
 
+mod accept;
 mod acceptor;
-mod serve;
 mod server;
 mod stream;
